+++ tt_metal/api/tt-metalium/dataflow_api.h
@@ -4,121 +4,12 @@
 
 #pragma once
 
-#if __has_include("chlkc_unpack_data_format.h")
-#include "chlkc_pack_data_format.h"
-#include "chlkc_unpack_data_format.h"
-#include "chlkc_unpack_tile_dims.h"
-#define DATA_FORMATS_DEFINED
-#endif
-#include <noc/noc_parameters.h>
-
-#include <stdint.h>
-
-#include "core_config.h"
-#include "circular_buffer.h"
-#include "dataflow_cmd_bufs.h"
-#include "debug/sanitize_noc.h"
-#include "debug/waypoint.h"
-#include "eth_l1_address_map.h"
-#include "hostdevcommon/common_values.hpp"
-#include "risc_attribs.h"
-#include "umd/device/tt_silicon_driver_common.hpp"
-#include "utils/utils.h"
-#include "debug/assert.h"
-#include "dev_msgs.h"
-
-#if defined(COMPILE_FOR_BRISC)
-constexpr uint8_t proc_type = static_cast<std::underlying_type_t<TensixProcessorTypes>>(TensixProcessorTypes::DM0);
-#else
-constexpr uint8_t proc_type = static_cast<std::underlying_type_t<TensixProcessorTypes>>(TensixProcessorTypes::DM1);
-#endif
-#if defined(KERNEL_BUILD)
-constexpr uint8_t noc_index = NOC_INDEX;
-constexpr uint8_t noc_mode = NOC_MODE;
-#else
-
-extern uint8_t noc_index;
-constexpr uint8_t noc_mode = DM_DEDICATED_NOC;
-#endif
-extern uint16_t dram_bank_to_noc_xy[NUM_NOCS][NUM_DRAM_BANKS];
-extern int32_t bank_to_dram_offset[NUM_DRAM_BANKS];
-extern uint16_t l1_bank_to_noc_xy[NUM_NOCS][NUM_L1_BANKS];
-extern int32_t bank_to_l1_offset[NUM_L1_BANKS];
-
-extern uint32_t tt_l1_ptr* rta_l1_base;
-extern uint32_t tt_l1_ptr* crta_l1_base;
-extern uint32_t tt_l1_ptr* sem_l1_base[];
-
-/** @file */
+#include "dataflow_api_declarations.h"
+#include "dataflow_addr_gen.h"
 
 /**
  * \private
  */
-
-// Use VC 1 for unicast writes, and VC 4 for mcast writes
-#define NOC_UNICAST_WRITE_VC 1
-#define NOC_MULTICAST_WRITE_VC 4
-#define NOC_DISPATCH_MULTICAST_WRITE_VC 5  // Only to be used by the dispatch cores
-
-#define EXCLUDE_ENABLED 1
-#define EXCLUDE_ENABLED_OFFSET 22
-#define EXCLUDE_DIRECTION_Y_OFFSET 21
-#define EXCLUDE_DIRECTION_X_OFFSET 20
-#define EXCLUDE_START_Y_OFFSET 14
-#define EXCLUDE_START_X_OFFSET 8
-#define DYNAMIC_NOC_DIRECTION(noc, direction) (noc == 1 ? 1 - direction : direction)
-
-static_assert(NUM_NOCS == 2);
-// "Scratch" in L1 has space allocated for 256 DRAM and L1 enteries, to store offsets and NOC XY data. (MEM_BANK_TO_NOC_XY_SCRATCH and MEM_BANK_OFFSET_SCRATCH)
-static_assert((NUM_DRAM_BANKS + NUM_L1_BANKS) <= 256);
-
-namespace interleaved_addr_gen {
-
-template <bool DRAM>
-FORCE_INLINE uint32_t get_bank_offset_index(uint32_t id) {
-    if constexpr (DRAM) {  // DRAM
-#ifdef IS_NOT_POW2_NUM_DRAM_BANKS
-        return udivsi3_const_divisor<NUM_DRAM_BANKS>(id);
-#else
-        return id >> LOG_BASE_2_OF_NUM_DRAM_BANKS;
-#endif
-    } else {  // L1
-#ifdef IS_NOT_POW2_NUM_L1_BANKS
-        return udivsi3_const_divisor<NUM_L1_BANKS>(id);
-#else
-        return id >> LOG_BASE_2_OF_NUM_L1_BANKS;
-#endif
-    }
-}
-
-template <bool DRAM>
-FORCE_INLINE uint32_t get_bank_index(uint32_t id, uint32_t bank_offset_index) {
-    if constexpr (DRAM) {  // DRAM
-        return id - bank_offset_index * NUM_DRAM_BANKS;
-    } else {  // L1
-        return id - bank_offset_index * NUM_L1_BANKS;
-    }
-}
-
-template <bool DRAM>
-FORCE_INLINE uint32_t get_noc_xy(uint32_t bank_index, uint8_t noc = noc_index) {
-    if constexpr (DRAM) {  // DRAM
-        return dram_bank_to_noc_xy[noc][bank_index];
-    } else {  // L1
-        return l1_bank_to_noc_xy[noc][bank_index];
-    }
-}
-
-template <bool DRAM>
-FORCE_INLINE uint32_t get_bank_offset(uint32_t bank_index) {
-    if constexpr (DRAM) {  // DRAM
-        return bank_to_dram_offset[bank_index];
-    } else {  // L1
-        return bank_to_l1_offset[bank_index];
-    }
-}
-
-}  // namespace interleaved_addr_gen
 
 /**
  * Returns the address in L1 for a given runtime argument index for unique (per core) runtime arguments set via
@@ -152,25 +43,6 @@ static FORCE_INLINE uint32_t get_arg_addr(int arg_idx) {
 static FORCE_INLINE uint32_t get_common_arg_addr(int arg_idx) { return (uint32_t)&crta_l1_base[arg_idx]; }
 
 /**
- * Returns the value at a given runtime argument index for unique (per-core) runtime arguments set via SetRuntimeArgs()
- * API.
- *
- * Return value: The value associated with the unique runtime argument index
- *
- * | Argument              | Description                                    | Type                  | Valid Range |
- * Required |
- * |-----------------------|------------------------------------------------|-----------------------|---------------------------|----------|
- * | arg_idx               | Unique Runtime argument index                  | uint32_t              | 0 to 255 | True |
- * | T (template argument) | Data type of the returned argument             | Any 4-byte sized type | N/A | True     |
- */
-template <typename T>
-FORCE_INLINE T get_arg_val(int arg_idx) {
-    // only 4B args are supported (eg int32, uint32)
-    static_assert("Error: only 4B args are supported" && sizeof(T) == 4);
-    return *((tt_l1_ptr T*)(get_arg_addr(arg_idx)));
-}
-
-/**
  * Returns the value at a given runtime argument index for common (all cores) runtime arguments set via
  * SetCommonRuntimeArgs() API.
  *
@@ -200,44 +72,6 @@ FORCE_INLINE T get_common_arg_val(int arg_idx) {
  * | arg_idx               | The index of the argument          | uint32_t              | 0 to 31     | True     |
  */
 #define get_compile_time_arg_val(arg_idx) KERNEL_COMPILE_TIME_ARG_##arg_idx
-
-template <uint32_t tile_hw = 1024>
-FORCE_INLINE constexpr static std::uint32_t MUL_WITH_TILE_SIZE(uint format, uint index) {
-    constexpr uint8_t datum_shift = (tile_hw == 1024)  ? 10
-                                    : (tile_hw == 512) ? 9
-                                    : (tile_hw == 256) ? 8
-                                    : (tile_hw == 128) ? 7
-                                    : (tile_hw == 64)  ? 6
-                                    : (tile_hw == 32)  ? 5
-                                    : (tile_hw == 16)  ? 4
-                                                       : 10;
-
-    constexpr uint8_t exp_shift = (tile_hw == 1024)  ? 6
-                                  : (tile_hw == 512) ? 5
-                                  : (tile_hw == 256) ? 4
-                                  : (tile_hw == 128) ? 4
-                                  : (tile_hw == 64)  ? 4
-                                  : (tile_hw == 32)  ? 4
-                                  : (tile_hw == 16)  ? 4
-                                                     : 6;
-    switch (format & 0x1F) {
-        case ((uint8_t)DataFormat::UInt8): return (index << datum_shift);
-        case ((uint8_t)DataFormat::UInt16):
-        case ((uint8_t)DataFormat::Float16):
-        case ((uint8_t)DataFormat::Float16_b): return (index << (datum_shift + 1));
-        case ((uint8_t)DataFormat::Int32):
-        case ((uint8_t)DataFormat::UInt32):
-        case ((uint8_t)DataFormat::Float32): return (index << (datum_shift + 2));
-        case ((uint8_t)DataFormat::Bfp2):
-        case ((uint8_t)DataFormat::Bfp2_b): return ((index << (datum_shift - 2)) + (index << (exp_shift)));
-        case ((uint8_t)DataFormat::Bfp4):
-        case ((uint8_t)DataFormat::Bfp4_b): return ((index << (datum_shift - 1)) + (index << (exp_shift)));
-        case ((uint8_t)DataFormat::Bfp8):
-        case ((uint8_t)DataFormat::Bfp8_b):
-        // Keep default as Bfp8?
-        default: return ((index << datum_shift) + (index << (exp_shift)));
-    };
-}
 
 /**
  * Pushes a given number of tiles in the back of the specified CB’s queue.
@@ -580,19 +414,6 @@ std::uint64_t get_noc_addr(std::uint32_t noc_x, std::uint32_t noc_y, std::uint32
     return NOC_XY_ADDR(DYNAMIC_NOC_X(noc, noc_x), DYNAMIC_NOC_Y(noc, noc_y), addr);
 }
 
-/*
-    Need an alias to get_noc_addr so that the structs below don't confuse the above get_noc_addr with
-    the struct variant
-*/
-FORCE_INLINE
-std::uint64_t get_noc_addr_helper(std::uint32_t noc_xy, std::uint32_t addr) {
-    /*
-        Get an encoding which contains tensix core and address you want to
-        write to via the noc multicast
-    */
-    return ((uint64_t)(noc_xy) << NOC_ADDR_COORD_SHIFT) | addr;
-}
-
 FORCE_INLINE
 std::uint32_t get_noc_exclude_region(
     std::uint32_t exclude_start_x,
@@ -661,74 +482,6 @@ std::uint64_t get_noc_addr(std::uint32_t addr, uint8_t noc = noc_index) {
         read from/write to via the noc
     */
     return NOC_XY_ADDR(my_x[noc], my_y[noc], addr);
-}
-
-// TODO: write docs
-// this issues only a single packet with size <= NOC_MAX_BURST_SIZE (ie maximum packet size)
-FORCE_INLINE
-void noc_async_read_one_packet(
-    std::uint64_t src_noc_addr, std::uint32_t dst_local_l1_addr, std::uint32_t size, uint8_t noc = noc_index) {
-    /*
-        Read requests - use static VC
-        Read responses - assigned VCs dynamically
-    */
-
-    WAYPOINT("RP2W");
-    while (!noc_cmd_buf_ready(noc, read_cmd_buf));
-    WAYPOINT("RP2D");
-
-    WAYPOINT("NAOW");
-    DEBUG_SANITIZE_NOC_READ_TRANSACTION(noc, src_noc_addr, dst_local_l1_addr, size);
-
-    NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_RET_ADDR_LO, dst_local_l1_addr);
-    NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_LO, (uint32_t)src_noc_addr);
-#ifdef ARCH_BLACKHOLE
-    // Handles reading from PCIe
-    NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_MID, (uint32_t)(src_noc_addr >> 32) & 0x1000000F);
-#endif
-    NOC_CMD_BUF_WRITE_REG(
-        noc,
-        read_cmd_buf,
-        NOC_TARG_ADDR_COORDINATE,
-        (uint32_t)(src_noc_addr >> NOC_ADDR_COORD_SHIFT) & NOC_COORDINATE_MASK);
-    NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_AT_LEN_BE, size);
-    NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-    noc_reads_num_issued[noc] += 1;
-
-    WAYPOINT("NAOD");
-}
-
-/**
- * Initiates an asynchronous read from a specified source node located at NOC
- * coordinates (x,y) at a local address (encoded as a uint64_t using \a
- * get_noc_addr function). The destination is in L1 memory on the Tensix core
- * executing this function call. Also, see \a noc_async_read_barrier.
- *
- * The source node can be either a DRAM bank, a Tensix core or a PCIe controller.
- *
- * Return value: None
- *
- * | Argument          | Description                                        | Data type | Valid range | required |
- * |-------------------|----------------------------------------------------|-----------|------------------------------------------|----------|
- * | src_noc_addr      | Encoding of the source NOC location (x,y)+address  | uint64_t  | DOX-TODO(ref to explain valid
- * coords)    | Yes      | | dst_local_l1_addr | Address in local L1 memory                         | uint32_t  | 0..1MB
- * | Yes      | | size              | Size of data transfer in bytes                     | uint32_t  | 0..1MB | Yes |
- */
-template <uint32_t max_page_size = NOC_MAX_BURST_SIZE + 1>
-inline void noc_async_read(
-    std::uint64_t src_noc_addr, std::uint32_t dst_local_l1_addr, std::uint32_t size, uint8_t noc = noc_index) {
-    /*
-        Read requests - use static VC
-        Read responses - assigned VCs dynamically
-    */
-    if constexpr (max_page_size <= NOC_MAX_BURST_SIZE) {
-        noc_async_read_one_packet(src_noc_addr, dst_local_l1_addr, size, noc);
-    } else {
-        WAYPOINT("NARW");
-        DEBUG_SANITIZE_NOC_READ_TRANSACTION(noc, src_noc_addr, dst_local_l1_addr, size);
-        ncrisc_noc_fast_read_any_len(noc, read_cmd_buf, src_noc_addr, dst_local_l1_addr, size);
-        WAYPOINT("NARD");
-    }
 }
 
 // TODO: write docs
@@ -1001,369 +754,6 @@ FORCE_INLINE void noc_async_write_one_packet_with_state(
     }
 }
 
-template <bool DRAM>
-struct InterleavedAddrGen {
-    uint32_t bank_base_address;  // Base address for the whole tensor.
-    const uint32_t page_size;    // Num bytes in page.
-    const uint32_t aligned_page_size = align_power_of_2(page_size, ALLOCATOR_ALIGNMENT);
-
-    FORCE_INLINE
-    uint32_t get_addr(
-        const uint32_t id,
-        const uint32_t bank_offset_index,
-        const uint32_t bank_index,
-        const uint32_t offset = 0) const {
-        return (bank_offset_index * this->aligned_page_size) + this->bank_base_address + offset +
-               interleaved_addr_gen::get_bank_offset<DRAM>(bank_index);
-    }
-
-    FORCE_INLINE
-    std::uint64_t get_noc_addr(const uint32_t id, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        uint64_t noc_addr = get_noc_addr_helper(noc_xy, addr);
-        return noc_addr;
-    }
-
-    FORCE_INLINE
-    void noc_async_read_page(
-        const uint32_t id, const uint32_t dest_addr, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        noc_async_read(this->get_noc_addr(id, offset), dest_addr, page_size, noc);
-    }
-};
-
-template <bool DRAM>
-struct InterleavedPow2AddrGen {
-    const uint32_t bank_base_address;
-    const uint32_t log_base_2_of_page_size;  // WARNING: This struct is used for optimized get_noc_addr in which case
-                                             // you know that bank_unit_size is a power of 2
-    const uint32_t aligned_log_base_2_of_page_size = this->log_base_2_of_page_size > LOG_BASE_2_OF_ALLOCATOR_ALIGNMENT
-                                                         ? this->log_base_2_of_page_size
-                                                         : LOG_BASE_2_OF_ALLOCATOR_ALIGNMENT;
-
-    FORCE_INLINE
-    uint32_t get_addr(
-        const uint32_t id,
-        const uint32_t bank_offset_index,
-        const uint32_t bank_index,
-        const uint32_t offset = 0) const {
-        return (bank_offset_index << this->aligned_log_base_2_of_page_size) + this->bank_base_address + offset +
-               interleaved_addr_gen::get_bank_offset<DRAM>(bank_index);
-    }
-
-    FORCE_INLINE
-    std::uint64_t get_noc_addr(const uint32_t id, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        uint64_t noc_addr = get_noc_addr_helper(noc_xy, addr);
-        return noc_addr;
-    }
-};
-
-template <bool DRAM, uint32_t tile_hw = 1024>
-struct InterleavedAddrGenFast {
-    uint32_t bank_base_address;  // Base address for the whole tensor.
-    // TODO: Remove page_size from argument list. This can be derived from data_format
-    uint32_t page_size;      // Num bytes in bank unit.
-    DataFormat data_format;  // Data format
-
-    FORCE_INLINE
-    uint32_t get_addr(
-        const uint32_t id,
-        const uint32_t bank_offset_index,
-        const uint32_t bank_index,
-        const uint32_t offset = 0) const {
-        return MUL_WITH_TILE_SIZE<tile_hw>((uint)this->data_format, bank_offset_index) + this->bank_base_address +
-               offset + interleaved_addr_gen::get_bank_offset<DRAM>(bank_index);
-    }
-
-    FORCE_INLINE
-    std::uint64_t get_noc_addr(const uint32_t id, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        uint64_t noc_addr = get_noc_addr_helper(noc_xy, addr);
-        return noc_addr;
-    }
-
-    FORCE_INLINE
-    void noc_async_read_tile(
-        const uint32_t id, uint32_t dest_addr, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t src_addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t src_noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        WAYPOINT("NRTW");
-        DEBUG_SANITIZE_NOC_READ_TRANSACTION(noc, get_noc_addr_helper(src_noc_xy, src_addr), dest_addr, this->page_size);
-        while (!noc_cmd_buf_ready(noc, read_cmd_buf));
-        WAYPOINT("NRTD");
-
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_RET_ADDR_LO, dest_addr);
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_LO, src_addr);            // (uint32_t)src_addr
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_COORDINATE, src_noc_xy);  // src_addr >> 32
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_AT_LEN_BE, this->page_size);        // len_bytes
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-        noc_reads_num_issued[noc] += 1;
-    }
-
-    FORCE_INLINE
-    void noc_async_write_tile(const uint32_t id, uint32_t src_addr, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t dest_addr = this->get_addr(id, bank_offset_index, bank_index);
-        uint32_t dest_noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        WAYPOINT("NWTW");
-        DEBUG_SANITIZE_NOC_WRITE_TRANSACTION(
-            noc, get_noc_addr_helper(dest_noc_xy, dest_addr), src_addr, this->page_size);
-        while (!noc_cmd_buf_ready(noc, write_cmd_buf));
-        WAYPOINT("NWTD");
-
-        constexpr uint32_t noc_cmd_field = NOC_CMD_CPY | NOC_CMD_WR | NOC_CMD_VC_STATIC |
-                                           NOC_CMD_STATIC_VC(NOC_UNICAST_WRITE_VC) |
-                                           0x0 |  // (linked ? NOC_CMD_VC_LINKED : 0x0)
-                                           0x0 |  // (mcast ? (NOC_CMD_PATH_RESERVE | NOC_CMD_BRCST_PACKET) : 0x0)
-                                           NOC_CMD_RESP_MARKED;
-
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_CTRL, noc_cmd_field);
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_TARG_ADDR_LO, src_addr);
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_RET_ADDR_LO, dest_addr);            // (uint32_t)dest_addr
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_RET_ADDR_COORDINATE, dest_noc_xy);  // dest_addr >> 32
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_AT_LEN_BE, this->page_size);        // len_bytes
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-        if constexpr (noc_mode == DM_DYNAMIC_NOC) {
-            inc_noc_nonposted_writes_acked<proc_type>(noc);
-        } else {
-            noc_nonposted_writes_num_issued[noc] += 1;
-            noc_nonposted_writes_acked[noc] += 1;  // num_dests
-        }
-    }
-};
-
-// TODO: add noc_async_write_page
-// TODO: need static assert + host assert that page size <= 8192, hard constraint
-template <bool DRAM>
-struct InterleavedPow2AddrGenFast {
-    uint32_t bank_base_address;              // Base address for the whole tensor.
-    const uint32_t log_base_2_of_page_size;  // Num bytes in bank unit.
-    const uint32_t aligned_log_base_2_of_page_size = this->log_base_2_of_page_size > LOG_BASE_2_OF_ALLOCATOR_ALIGNMENT
-                                                         ? this->log_base_2_of_page_size
-                                                         : LOG_BASE_2_OF_ALLOCATOR_ALIGNMENT;
-
-    FORCE_INLINE
-    uint32_t get_addr(
-        const uint32_t id,
-        const uint32_t bank_offset_index,
-        const uint32_t bank_index,
-        const uint32_t offset = 0) const {
-        return (bank_offset_index << this->aligned_log_base_2_of_page_size) + this->bank_base_address + offset +
-               interleaved_addr_gen::get_bank_offset<DRAM>(bank_index);
-    }
-
-    FORCE_INLINE
-    std::uint64_t get_noc_addr(const uint32_t id, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        uint64_t noc_addr = get_noc_addr_helper(noc_xy, addr);
-        return noc_addr;
-    }
-
-    FORCE_INLINE
-    void noc_async_read_page(
-        const uint32_t id, uint32_t dest_addr, const uint32_t offset = 0, uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t src_addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t src_noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        WAYPOINT("NRPW");
-        DEBUG_SANITIZE_NOC_READ_TRANSACTION(
-            noc, get_noc_addr_helper(src_noc_xy, src_addr), dest_addr, 1 << this->aligned_log_base_2_of_page_size);
-        while (!noc_cmd_buf_ready(noc, read_cmd_buf));
-        WAYPOINT("NRPD");
-
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_RET_ADDR_LO, dest_addr);
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_LO, src_addr);            // (uint32_t)src_addr
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_COORDINATE, src_noc_xy);  // src_addr >> 32
-        NOC_CMD_BUF_WRITE_REG(
-            noc, read_cmd_buf, NOC_AT_LEN_BE, 1 << this->aligned_log_base_2_of_page_size);  // len_bytes
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-        noc_reads_num_issued[noc] += 1;
-    }
-
-    FORCE_INLINE
-    void noc_async_read_partial_page(
-        const uint32_t id,
-        uint32_t dest_addr,
-        const uint32_t size,
-        const uint32_t offset,
-        uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t src_addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t src_noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        WAYPOINT("RP1W");
-        while (!noc_cmd_buf_ready(noc, read_cmd_buf));
-        WAYPOINT("RP1D");
-        DEBUG_SANITIZE_NOC_READ_TRANSACTION(noc, get_noc_addr_helper(src_noc_xy, src_addr), dest_addr, size);
-
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_RET_ADDR_LO, dest_addr);
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_LO, src_addr);            // (uint32_t)src_addr
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_TARG_ADDR_COORDINATE, src_noc_xy);  // src_addr >> 32
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_AT_LEN_BE, size);                   // len_bytes
-        NOC_CMD_BUF_WRITE_REG(noc, read_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-        noc_reads_num_issued[noc] += 1;
-    }
-
-    FORCE_INLINE
-    void noc_async_write_page(
-        const uint32_t id,
-        uint32_t src_addr,
-        const uint32_t write_size_bytes,
-        const uint32_t offset = 0,
-        uint8_t noc = noc_index) const {
-        uint32_t bank_offset_index = interleaved_addr_gen::get_bank_offset_index<DRAM>(id);
-        uint32_t bank_index = interleaved_addr_gen::get_bank_index<DRAM>(id, bank_offset_index);
-        uint32_t dest_addr = this->get_addr(id, bank_offset_index, bank_index, offset);
-        uint32_t dest_noc_xy = interleaved_addr_gen::get_noc_xy<DRAM>(bank_index, noc);
-
-        WAYPOINT("NWPW");
-        DEBUG_SANITIZE_NOC_WRITE_TRANSACTION(
-            noc, get_noc_addr_helper(dest_noc_xy, dest_addr), src_addr, write_size_bytes);
-        while (!noc_cmd_buf_ready(noc, write_cmd_buf));
-        WAYPOINT("NWPD");
-
-        constexpr uint32_t noc_cmd_field = NOC_CMD_CPY | NOC_CMD_WR | NOC_CMD_VC_STATIC |
-                                           NOC_CMD_STATIC_VC(NOC_UNICAST_WRITE_VC) |
-                                           0x0 |  // (linked ? NOC_CMD_VC_LINKED : 0x0)
-                                           0x0 |  // (mcast ? (NOC_CMD_PATH_RESERVE | NOC_CMD_BRCST_PACKET) : 0x0)
-                                           NOC_CMD_RESP_MARKED;
-
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_CTRL, noc_cmd_field);
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_TARG_ADDR_LO, src_addr);
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_RET_ADDR_LO, dest_addr);            // (uint32_t)dest_addr
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_RET_ADDR_COORDINATE, dest_noc_xy);  // dest_addr >> 32
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_AT_LEN_BE, write_size_bytes);       // len_bytes
-        NOC_CMD_BUF_WRITE_REG(noc, write_cmd_buf, NOC_CMD_CTRL, NOC_CTRL_SEND_REQ);
-        if constexpr (noc_mode == DM_DYNAMIC_NOC) {
-            inc_noc_nonposted_writes_acked<proc_type>(noc);
-        } else {
-            noc_nonposted_writes_num_issued[noc] += 1;
-            noc_nonposted_writes_acked[noc] += 1;  // num_dests
-        }
-
-    }
-};
-
-template <bool DRAM>
-FORCE_INLINE std::uint64_t get_noc_addr(
-    const uint32_t id, const InterleavedAddrGen<DRAM>& s, uint32_t offset = 0, uint8_t noc = noc_index) {
-    /*
-        Alternative API for getting the noc address when we are reading using a swizzled
-        layout. This version assumes bank unit size can be arbitrary size. Use
-        get_noc_addr(const uint32_t id, InterleavedPow2AddrGen s) for optimized algorithm in which stick size
-        is a power of 2.
-
-        id: Unique id for the bank_unit you want to read, assuming row major order. We use this to compute the
-        bank for this unit of data.
-
-        InterleavedAddrGen: Check struct for attribute definitions.
-    */
-    return s.get_noc_addr(id, offset, noc);
-}
-
-template <bool DRAM>
-FORCE_INLINE std::uint64_t get_noc_addr(
-    const uint32_t id, const InterleavedPow2AddrGen<DRAM>& s, uint32_t offset = 0, uint8_t noc = noc_index) {
-    /*
-        Alternative API for getting the noc address when we are reading using a swizzled
-        layout. This version assumes bank unit size is a power of 2. For arbitrary bank
-        unit size, use get_noc_addr(const uint32_t id, const InterleavedOffset s)
-
-        id: Unique id for the bank_unit you want to read, assuming row major order. We use this to compute the
-        bank for this unit of data.
-
-        InterleavedPow2AddrGen: Check struct for attribute definitions.
-    */
-
-    return s.get_noc_addr(id, offset, noc);
-}
-
-template <bool DRAM, uint32_t tile_hw>
-FORCE_INLINE std::uint64_t get_noc_addr(
-    const uint32_t id, const InterleavedAddrGenFast<DRAM, tile_hw>& s, uint32_t offset = 0, uint8_t noc = noc_index) {
-    /*
-        Alternative API for getting the noc address when we are reading using a swizzled
-        layout. This version assumes bank unit size can be arbitrary size. Use
-        get_noc_addr(const uint32_t id, InterleavedPow2AddrGen s) for optimized algorithm in which stick size
-        is a power of 2.
-
-        id: Unique id for the bank_unit you want to read, assuming row major order. We use this to compute the
-        bank for this unit of data.
-
-        InterleavedAddrGen: Check struct for attribute definitions.
-    */
-    return s.get_noc_addr(id, offset, noc);
-}
-
-template <bool DRAM>
-FORCE_INLINE std::uint64_t get_noc_addr(
-    const uint32_t id, const InterleavedPow2AddrGenFast<DRAM>& s, uint32_t offset = 0, uint8_t noc = noc_index) {
-    /*
-        Alternative API for getting the noc address when we are reading using a swizzled
-        layout. This version assumes bank unit size is a power of 2 and less than or equal to NOC_MAX_BURST_SIZE.
-        For arbitrary bank unit size, use get_noc_addr(const uint32_t id, const InterleavedOffset s)
-
-        id: Unique id for the bank_unit you want to read, assuming row major order. We use this to compute the
-        bank for this unit of data.
-
-        InterleavedPow2AddrGenFast: Check struct for attribute definitions.
-    */
-
-    return s.get_noc_addr(id, offset, noc);
-}
-
-template <bool DRAM>
-FORCE_INLINE void noc_async_read_page(
-    const uint32_t id,
-    const InterleavedAddrGen<DRAM>& s,
-    std::uint32_t dst_local_l1_addr,
-    uint32_t offset = 0,
-    uint8_t noc = noc_index) {
-    /*
-        Read requests - use static VC
-        Read responses - assigned VCs dynamically
-    */
-    s.noc_async_read_page(id, dst_local_l1_addr, offset, noc);
-}
-
-template <bool DRAM, uint32_t tile_hw>
-FORCE_INLINE void noc_async_read_tile(
-    const uint32_t id,
-    const InterleavedAddrGenFast<DRAM, tile_hw>& s,
-    std::uint32_t dst_local_l1_addr,
-    uint32_t offset = 0,
-    uint8_t noc = noc_index) {
-    /*
-        Read requests - use static VC
-        Read responses - assigned VCs dynamically
-    */
-    s.noc_async_read_tile(id, dst_local_l1_addr, offset, noc);
-}
-
 /**
  * Initiates an asynchronous write from a source address in L1 memory on the
  * Tensix core executing this function call. The destination is specified using
@@ -1395,15 +785,6 @@ inline void noc_async_write(
             noc, write_cmd_buf, src_local_l1_addr, dst_noc_addr, size, NOC_UNICAST_WRITE_VC, false, false, 1, true);
         WAYPOINT("NAWD");
     }
-}
-
-template <bool DRAM, uint32_t tile_hw>
-FORCE_INLINE void noc_async_write_tile(
-    const uint32_t id,
-    const InterleavedAddrGenFast<DRAM, tile_hw>& s,
-    std::uint32_t src_local_l1_addr,
-    uint8_t noc = noc_index) {
-    s.noc_async_write_tile(id, src_local_l1_addr, noc);
 }
 
 template <ProgrammableCoreType type = ProgrammableCoreType::TENSIX>
@@ -2011,21 +1392,4 @@ uint64_t get_noc_addr_from_bank_id(uint32_t bank_id, uint32_t bank_address_offse
         noc_addr = l1_bank_to_noc_xy[noc_index][bank_id];
     }
     return (noc_addr << NOC_ADDR_COORD_SHIFT) | (bank_address_offset);
-}
-
-template <bool DRAM, uint32_t page_size>
-FORCE_INLINE auto get_interleaved_addr_gen(uint32_t base_addr) {
-    constexpr bool is_pow_2 = is_power_of_2(page_size);
-    if constexpr (is_pow_2) {
-        constexpr uint32_t log2_page_size = __builtin_ctz(page_size);
-        if constexpr (page_size <= NOC_MAX_BURST_SIZE) {
-            return InterleavedPow2AddrGenFast<DRAM>{
-                .bank_base_address = base_addr, .log_base_2_of_page_size = log2_page_size};
-        } else {
-            return InterleavedPow2AddrGen<DRAM>{
-                .bank_base_address = base_addr, .log_base_2_of_page_size = log2_page_size};
-        }
-    } else {
-        return InterleavedAddrGen<DRAM>{.bank_base_address = base_addr, .page_size = page_size};
-    }
 }
