+++ ttnn/cpp/ttnn/operations/ccl/common/host/ccl_worker_builder.cpp
@@ -876,7 +876,7 @@ KernelHandle generate_multi_command_stream_kernel_ct_args(
         }
         for (size_t i = 0; i < tensors.size(); i++) {
             std::ranges::copy(
-                ttnn::ccl::emit_address_generator_compile_time_args(*tensors[i]), std::back_inserter(ct_args));
+                ttnn::ccl::legacy_emit_address_generator_compile_time_args(*tensors[i]), std::back_inserter(ct_args));
         }
 
         datamovement_kernel_config.compile_args = ct_args;
@@ -1131,11 +1131,11 @@ void generate_multi_input_command_stream_kernel_rt_args(
             if (tensor_device_override.has_value() and
                 tensor_device_override.value().find(t) != tensor_device_override.value().end()) {
                 std::ranges::copy(
-                    ttnn::ccl::emit_address_generator_runtime_args(tensor_device_override->at(t), *t),
+                    ttnn::ccl::legacy_emit_address_generator_runtime_args(tensor_device_override->at(t), *t),
                     std::back_inserter(rt_args));
             } else {
                 std::ranges::copy(
-                    ttnn::ccl::emit_address_generator_runtime_args(t->buffer()->device(), *t),
+                    ttnn::ccl::legacy_emit_address_generator_runtime_args(t->buffer()->device(), *t),
                     std::back_inserter(rt_args));
             }
         } else {
@@ -1225,7 +1225,7 @@ void generate_multi_command_stream_kernel_rt_args(
 
     for (size_t i = 0; i < num_command_streams; i++) {
         std::ranges::copy(
-            ttnn::ccl::emit_address_generator_runtime_args(device, *tensors[i]), std::back_inserter(rt_args));
+            ttnn::ccl::legacy_(device, *tensors[i]), std::back_inserter(rt_args));
     }
 
     // TODO: Handle teardown signalling
@@ -1441,7 +1441,7 @@ std::vector<uint32_t> CCLWorkerArgBuilder::generate_sender_reader_kernel_rt_args
     log_trace(tt::LogOp, "ccl_send_reader arg[{}]: page_size {}", logged_arg_idx, args[logged_arg_idx]);
     logged_arg_idx++;
 
-    auto const& addr_gen_rt_args = ttnn::ccl::emit_address_generator_runtime_args(this->device, input_tensor);
+    auto const& addr_gen_rt_args = ttnn::ccl::legacy_emit_address_generator_runtime_args(this->device, input_tensor);
     std::ranges::copy(addr_gen_rt_args, std::back_inserter(args));
     for (auto const& arg : addr_gen_rt_args) {
         log_trace(tt::LogOp, "ccl_send_reader arg[{}]: addr_gen_rt_args[] {}", logged_arg_idx, args[logged_arg_idx]);
@@ -1598,7 +1598,7 @@ std::vector<uint32_t> CCLWorkerArgBuilder::generate_sender_writer_kernel_rt_args
         }
     }
 
-    auto const& addr_gen_rt_args = ttnn::ccl::emit_address_generator_runtime_args(this->device, output_tensor);
+    auto const& addr_gen_rt_args = ttnn::ccl::legacy_emit_address_generator_runtime_args(this->device, output_tensor);
     std::ranges::copy(addr_gen_rt_args, std::back_inserter(args));
     for (auto const& arg : addr_gen_rt_args) {
         log_trace(tt::LogOp, "ccl_send_writer arg[{}]: addr_gen_rt_args[] {}", logged_arg_idx, args[logged_arg_idx]);
@@ -1622,7 +1622,7 @@ std::vector<uint32_t> CCLWorkerArgBuilder::generate_sender_reader_kernel_ct_args
         static_cast<uint32_t>(tt::CB::c_in0)                                // cb_id
     };
 
-    auto const& addr_gen_rt_args = ttnn::ccl::emit_address_generator_compile_time_args(input_tensor);
+    auto const& addr_gen_rt_args = ttnn::ccl::legacy_emit_address_generator_compile_time_args(input_tensor);
     std::ranges::copy(addr_gen_rt_args, std::back_inserter(args));
 
     return args;
@@ -1637,7 +1637,7 @@ std::vector<uint32_t> CCLWorkerArgBuilder::generate_sender_writer_kernel_ct_args
         static_cast<uint32_t>(tt::CB::c_in0)                                 // cb_id
     };
 
-    auto const& addr_gen_rt_args = ttnn::ccl::emit_address_generator_compile_time_args(output_tensor);
+    auto const& addr_gen_rt_args = ttnn::ccl::legacy_emit_address_generator_compile_time_args(output_tensor);
     std::ranges::copy(addr_gen_rt_args, std::back_inserter(args));
 
     return args;
